flacenc::bitsink
Struct MemSink

pub struct MemSink<S> { /* private fields */ }

BitSink implementation based on Vec of unsigned ints (u8/ u64).

Currently, only S: u64 and S: u8 implements BitSink trait.

Implementations

impl<S: Bits> MemSink<S>

pub fn new() -> Self

Creates new MemSink instance with the default capacity.

Examples

let mut sink = MemSink::<u8>::new();
let empty: [u8; 0] = [];
assert_eq!(&empty, sink.as_slice());

pub fn with_capacity(capacity_in_bits: usize) -> Self

Creates new MemSink instance with the specified capacity (in bits).

Examples

let mut sink = MemSink::<u8>::with_capacity(128);
sink.write_lsbs(0x00FFu16, 10);
assert!(sink.into_inner().capacity() > 128 / 8);

pub fn clear(&mut self)

Clears the vector, removing all values.

Examples

let mut sink = MemSink::<u8>::new();
sink.write_lsbs(0xAAAAAAAAu32, 14);
assert_eq!(sink.to_bitstring(), "10101010_101010**");
sink.clear();
assert_eq!(sink.to_bitstring(), "");

pub fn len(&self) -> usize

Returns the number of bits stored in the buffer.

Examples

let mut sink = MemSink::<u8>::new();
sink.write(0u64);
sink.write_msbs(0u8, 6);
assert_eq!(sink.len(), 70)

pub fn is_empty(&self) -> bool

Checks if the buffer is empty.

Examples

let mut sink = MemSink::<u8>::new();
assert!(sink.is_empty());
sink.write_msbs(0u8, 6);
assert!(!sink.is_empty());

pub fn reserve(&mut self, additional_in_bits: usize)

Reseerves capacity for at least additional_in_bits more bits.

This function reserves the ’Vec’s capacity so that the allocated size is sufficient for storing self.len() + additional_in_bits bits.

Examples

let mut sink = MemSink::<u8>::with_capacity(1);
sink.write_bytes_aligned(&[0u8; 128]);
assert_eq!(sink.len(), 1024);
sink.reserve(2048);
assert!(sink.into_inner().capacity() > (1024 + 2048) / 8);

pub fn to_bitstring(&self) -> String

Returns bits in a string.

This function formats an internal buffer state to a human-readable string. Each byte is shown in eight characters joined by '_', and the last bits of the last byte that are not yet filled are shown as '*'.

Examples

let mut sink = MemSink::<u8>::new();
sink.write_msbs(0x3456u16, 13);
assert_eq!(sink.to_bitstring(), "00110100_01010***");

pub fn into_inner(self) -> Vec<S>

Consumes ByteSink and returns the internal buffer.

Examples

let mut sink = MemSink::<u8>::new();
sink.write_bytes_aligned(&[0xABu8; 4]);
let v: Vec<u8> = sink.into_inner();
assert_eq!(&v, &[0xAB; 4]);

pub fn as_slice(&self) -> &[S]

Returns a reference to the internal bytes.

Examples

let mut sink = MemSink::<u8>::new();
sink.write_msbs(0x3456u16, 13);
assert_eq!(sink.as_slice(), &[0x34, 0x50]);

pub fn write_to_byte_slice(&self, dest: &mut [u8])

Writes the contents to a mutable byte slice.

For MemSink<u8>, this function basically makes a copy of the inner storage. For other instances, this function reorders and flattens the byte structure of the inner storage and copies bytes to the given mutable slice.

Examples

let mut sink = MemSink::<u64>::new();
sink.write_msbs(0xCAFE_FEED_BEEF_FACEu64, 47);
let mut bytes = [0u8; 6];
sink.write_to_byte_slice(&mut bytes);

// The last bit of the last byte (that is not written) is padded with 0.
assert_eq!(bytes,
           [0xCA, 0xFE, 0xFE, 0xED, 0xBE, 0xEE]);