flacenc::bitsink
Trait BitSink

pub trait BitSink: Sized {
    type Error: Error;

    // Required methods
    fn align_to_byte(&mut self) -> Result<usize, Self::Error>;
    fn write_lsbs<T: Bits>(
        &mut self,
        val: T,
        n: usize,
    ) -> Result<(), Self::Error>;
    fn write_msbs<T: Bits>(
        &mut self,
        val: T,
        n: usize,
    ) -> Result<(), Self::Error>;
    fn write<T: Bits>(&mut self, val: T) -> Result<(), Self::Error>;

    // Provided methods
    fn write_bytes_aligned(
        &mut self,
        bytes: &[u8],
    ) -> Result<usize, Self::Error> { ... }
    fn write_twoc<T: SignedBits>(
        &mut self,
        val: T,
        bits_per_sample: usize,
    ) -> Result<(), Self::Error> { ... }
    fn write_zeros(&mut self, n: usize) -> Result<(), Self::Error> { ... }
}

Storage-agnostic interface trait for bit-based output.

The encoder repeatedly generates arrays of code bits that are typically smaller than a byte (8 bits). A type implementing BitSink is used to arrange those bits typically in bytes, and transfer them to the backend storage. ByteSink (alias to MemSink<u8>) is a standard implementation of BitSink that stores code bits to a Vec of u8s.

Required Associated Types

type Error: Error

Error type that may happen while writing bits to BitSink.

Required Methods

fn align_to_byte(&mut self) -> Result<usize, Self::Error>

Puts zeros to BitSink until the length aligns to the byte boundaries.

Returns

The number of zeros put.

Errors

It can emit errors describing backend issues.

Examples

use flacenc::bitsink::{ByteSink, BitSink};
let mut sink = ByteSink::new();

sink.write_lsbs(0xFFu8, 3);
assert_eq!(sink.len(), 3);

let pads = sink.align_to_byte()?;
assert_eq!(pads, 5);
assert_eq!(sink.len(), 8);

fn write_lsbs<T: Bits>(&mut self, val: T, n: usize) -> Result<(), Self::Error>

Writes n LSBs to the sink.

Errors

It can emit errors describing backend issues.

Examples

use flacenc::bitsink::{ByteSink, BitSink};

let mut sink = ByteSink::new();
sink.write_lsbs(0x0Fu8, 3);

assert_eq!(sink.len(), 3);
assert_eq!(sink.to_bitstring(), "111*****");

fn write_msbs<T: Bits>(&mut self, val: T, n: usize) -> Result<(), Self::Error>

Writes n MSBs to the sink.

Errors

It can emit errors describing backend issues.

Examples

use flacenc::bitsink::{ByteSink, BitSink};

let mut sink = ByteSink::new();
sink.write_msbs(0xF0u8, 3);

assert_eq!(sink.to_bitstring(), "111*****");

fn write<T: Bits>(&mut self, val: T) -> Result<(), Self::Error>

Writes all bits in val: Bits.

Errors

It can emit errors describing backend issues.

Examples

use flacenc::bitsink::{ByteSink, BitSink};

let mut sink = ByteSink::new();
sink.write_msbs(0xF0u8, 3);

sink.write(0x5555u16);

assert_eq!(sink.to_bitstring(), "11101010_10101010_101*****");

Provided Methods

fn write_bytes_aligned(&mut self, bytes: &[u8]) -> Result<usize, Self::Error>

Writes bytes after alignment, and returns padded bits.

Errors

It can emit errors describing backend issues.

Examples

let mut sink = ByteSink::new();

sink.write_lsbs(0xFFu8, 3);
assert_eq!(sink.len(), 3);

sink.write_bytes_aligned(&[0xB7, 0x7D])?;

assert_eq!(sink.to_bitstring(), "11100000_10110111_01111101");

Even when we use u64 as a backend storage, this function should move the cursor on the eight-bit boundary.

let mut sink = MemSink::<u64>::new();

sink.write_lsbs(0xFFu8, 3);

sink.write_bytes_aligned(&[0xAA, 0xF0])?;

assert_eq!(
    sink.to_bitstring(),
    "111000001010101011110000****************************************"
);

fn write_twoc<T: SignedBits>( &mut self, val: T, bits_per_sample: usize, ) -> Result<(), Self::Error>

Writes val in two’s coplement format.

Errors

It can emit errors describing backend issues.

Examples

use flacenc::bitsink::{ByteSink, BitSink};

let mut sink = ByteSink::new();
sink.write_msbs(0xF0u8, 3);
assert_eq!(sink.to_bitstring(), "111*****");

// two's complement of 00011 in 11101
sink.write_twoc(-3i32, 5);
assert_eq!(sink.to_bitstring(), "11111101");

fn write_zeros(&mut self, n: usize) -> Result<(), Self::Error>

Writes n-bits of zeros.

A default implementation using write_msbs is provided. An impl can provide a faster short-cut for writing zeros.

Errors

It can emit errors describing backend issues.

Examples

use flacenc::bitsink::{ByteSink, BitSink};

let mut sink = ByteSink::new();
sink.write_msbs(0xF0u8, 3);
assert_eq!(sink.to_bitstring(), "111*****");

sink.write_zeros(6);
assert_eq!(sink.to_bitstring(), "11100000_0*******");

use flacenc::bitsink::{MemSink, BitSink};
let mut sink = MemSink::<u64>::new();
sink.write_lsbs(1u16, 1);
sink.write_zeros(65);
sink.write_lsbs(1u16, 1);

// a bitstring of `MemSink<u64>` has chunks with the lengths=64.
assert_eq!(
    sink.to_bitstring(),
    concat!(
        "1000000000000000000000000000000000000000000000000000000000000000_",
        "001*************************************************************",
    )
);

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.