flacenc::component
Struct StreamInfo

pub struct StreamInfo { /* private fields */ }

METADATA_BLOCK_STREAM_INFO component.

Implementations

impl StreamInfo

pub fn new( sample_rate: usize, channels: usize, bits_per_sample: usize, ) -> Result<Self, VerifyError>

Constructs new StreamInfo.

For unspecified fields, the following default values are used:

min_block_size: u16::MAX,

max_block_size: 0,

min_frame_size: u32::MAX,

max_frame_size: 0,

total_samples: 0,

md5_digest: [0u8; 16] (indicating verification disabled.)

Errors

Returns an error if an input argument is out of range.

Examples

let info = StreamInfo::new(16000, 2, 16).unwrap();
assert_eq!(info.max_frame_size(), 0);

pub fn update_frame_info(&mut self, frame: &Frame)

Updates StreamInfo with values from the given Frame.

This function updates {min|max}_{block|frame}_size and total_samples.

Examples

let (signal_len, block_size, channels, sample_rate) = (31234, 160, 2, 16000);
let other_stream = make_example_stream(signal_len, block_size, channels, sample_rate);
let mut info = StreamInfo::new(16000, 2, 16).unwrap();

for n in 0..other_stream.frame_count() {
    info.update_frame_info(other_stream.frame(n).unwrap());
}
assert_eq!(info.max_block_size(), 160);
assert_eq!(info.min_block_size(), 160);

// `Frame` doesn't hold the original signal length, so `total_samples`
// becomes a multuple of block_size == 160.
assert_eq!(info.total_samples(), 31360);

pub fn min_frame_size(&self) -> usize

Returns min_frame_size field.

Examples

let mut info = StreamInfo::new(16000, 2, 16).unwrap();

let (signal_len, block_size, channels, sample_rate) = (31234, 160, 2, 16000);
let other_stream = make_example_stream(signal_len, block_size, channels, sample_rate);
info.update_frame_info(other_stream.frame(0).unwrap());

assert!(info.min_frame_size() > 0);

pub fn max_frame_size(&self) -> usize

Returns max_frame_size field.

Examples

let info = StreamInfo::new(16000, 2, 16).unwrap();

assert_eq!(info.max_frame_size(), 0);

pub fn min_block_size(&self) -> usize

Returns min_block_size field.

Examples

let mut info = StreamInfo::new(16000, 2, 16).unwrap();

let (signal_len, block_size, channels, sample_rate) = (31234, 160, 2, 16000);
let other_stream = make_example_stream(signal_len, block_size, channels, sample_rate);
info.update_frame_info(other_stream.frame(0).unwrap());

assert_eq!(info.min_block_size(), 160);

pub fn max_block_size(&self) -> usize

Returns max_block_size field.

Examples

let info = StreamInfo::new(16000, 2, 16).unwrap();

assert_eq!(info.max_block_size(), 0);

pub fn sample_rate(&self) -> usize

Returns sample_rate field.

Examples

let info = StreamInfo::new(16000, 2, 16).unwrap();
assert_eq!(info.sample_rate(), 16000);

pub fn channels(&self) -> usize

Returns channels field.

Examples

let info = StreamInfo::new(16000, 2, 16).unwrap();
assert_eq!(info.channels(), 2);

pub fn bits_per_sample(&self) -> usize

Returns bits_per_sample field.

Examples

let info = StreamInfo::new(16000, 2, 16).unwrap();
assert_eq!(info.bits_per_sample(), 16);

pub fn total_samples(&self) -> usize

Returns total_samples field.

Examples

let (signal_len, block_size, channels, sample_rate) = (31234, 160, 2, 16000);
let other_stream = make_example_stream(signal_len, block_size, channels, sample_rate);
let mut info = StreamInfo::new(16000, 2, 16).unwrap();

for n in 0..other_stream.frame_count() {
    info.update_frame_info(other_stream.frame(n).unwrap());
}

// `Frame` doesn't hold the original signal length, so `total_samples`
// becomes a multuple of block_size == 160.
assert_eq!(info.total_samples(), 31360);

pub fn set_total_samples(&mut self, n: usize)

Sets total_samples field.

total_samples is updated during the encoding. However, since Frame only knows its frame size, the effective number of samples is not visible after paddings. Similar to set_md5_digest, this field should be finalized by propagating information from Context.

Examples

let mut ctx = Context::new(16, 2, 123);
let mut info = StreamInfo::new(16000, 2, 16).unwrap();
ctx.fill_interleaved(&[0x0000_0FFFi32; 246]);
info.set_total_samples(ctx.total_samples());
assert_ne!(info.total_samples(), 246);

pub fn md5_digest(&self) -> &[u8; 16]

Returns md5_digest field.

Examples

let (signal_len, block_size, channels, sample_rate) = (31234, 160, 2, 16000);
let other_stream = make_example_stream(signal_len, block_size, channels, sample_rate);
let mut info = StreamInfo::new(16000, 2, 16).unwrap();

assert_eq!(info.md5_digest(), &[0u8; 16]); // default

for n in 0..other_stream.frame_count() {
    info.update_frame_info(other_stream.frame(n).unwrap());
}

// `update_frame_info` doesn't update MD5
assert_eq!(info.md5_digest(), &[0u8; 16]);

pub fn set_md5_digest(&mut self, digest: &[u8; 16])

Resets MD5 digest value by the given slice.

MD5 computation is not performed in in update_frame_info, and is expected to be done externally (by Context). This function is called to set MD5 bytes after we read all input samples.

Examples

let mut ctx = Context::new(16, 2, 128);
let mut info = StreamInfo::new(16000, 2, 16).unwrap();
assert_eq!(info.md5_digest(), &[0x00u8; 16]);
ctx.fill_interleaved(&[0x0000_0FFFi32; 256]);
info.set_md5_digest(&ctx.md5_digest());
assert_ne!(info.md5_digest(), &[0x00u8; 16]);

pub fn set_block_sizes( &mut self, min_value: usize, max_value: usize, ) -> Result<(), VerifyError>

Resets min_block_size and max_block_size fields.

Errors

Returns error when min_value or max_value is not a valid block size, or when min_value > max_value.

Examples

let mut info = StreamInfo::new(16000, 2, 16).unwrap();
info.set_block_sizes(128, 1024).unwrap();
assert_eq!(info.min_block_size(), 128);
assert_eq!(info.max_block_size(), 1024);

pub fn set_frame_sizes( &mut self, min_value: usize, max_value: usize, ) -> Result<(), VerifyError>

Resets min_frame_size and max_frame_size fields.

Errors

Returns error when min_value or max_value is not 32-bit representable, or when min_value > max_value.

Examples

let mut info = StreamInfo::new(16000, 2, 16).unwrap();
info.set_frame_sizes(128, 2048).unwrap();
assert_eq!(info.min_frame_size(), 128);
assert_eq!(info.max_frame_size(), 2048);