flacenc::component
Trait BitRepr

pub trait BitRepr: Sealed {
    // Required methods
    fn count_bits(&self) -> usize;
    fn write<S: BitSink>(&self, dest: &mut S) -> Result<(), OutputError<S>>;
}

FLAC components that can be represented in a bit sequence.

Required Methods

fn count_bits(&self) -> usize

Counts the number of bits required to store the component.

fn write<S: BitSink>(&self, dest: &mut S) -> Result<(), OutputError<S>>

Writes the bit sequence to BitSink.

Errors

This function returns error if self contains an invalid value that does not fit to FLAC’s bitstream format, or if a BitSink method returned an error.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.