flacenc::error
Trait Verify

pub trait Verify: Sized + Sealed {
    // Required method
    fn verify(&self) -> Result<(), VerifyError>;

    // Provided methods
    fn into_verified(self) -> Result<Verified<Self>, (Self, VerifyError)> { ... }
    unsafe fn assume_verified(self) -> Verified<Self> { ... }
}

Trait for verifiable structs.

Required Methods

fn verify(&self) -> Result<(), VerifyError>

Verifies there’s no internal data inconsistency.

Errors

Returns VerifyError if there’s an invalid variable.

Examples

config::Prc imlpements Verify.

let mut prc = Prc::default();
prc.max_parameter = 256;  // invalid setting
assert!(prc.verify().is_err());

prc.max_parameter = 10; // valid setting
assert!(prc.verify().is_ok());

Provided Methods

fn into_verified(self) -> Result<Verified<Self>, (Self, VerifyError)>

Wraps into Verified to indicate that the data is already verified.

Errors

Returns the original input and VerifyError if verify failed.

Examples

config::Prc imlpements Verify.

let mut prc = Prc::default();
fn do_something_with_prc_config(config: &Verified<Prc>) -> usize {
    if config.max_parameter > MAX_RICE_PARAMETER {
        panic!();
    }
    // do something
    config.max_parameter * 2
}

prc.max_parameter = 256;  // invalid setting
assert!(
    prc.clone()
       .into_verified()
       .map(|x| do_something_with_prc_config(&x))
       .is_err()
);  // at least this does not panic

prc.max_parameter = 7;
assert_eq!(
    prc.into_verified()
       .map(|x| do_something_with_prc_config(&x))
       .unwrap(),
    14,
);

unsafe fn assume_verified(self) -> Verified<Self>

Wraps into Verified without actual verification.

Safety

The use of Verified data obtained this way may cause an unexpected behavior. It should be okay if the data are previously verified with verify function and have not been changed after that.

Examples

config::Prc imlpements Verify.

fn do_something_with_prc_config(config: &Verified<Prc>) -> usize {
    if config.max_parameter > MAX_RICE_PARAMETER {
        panic!();
    }
    // do something
    config.max_parameter * 2
}
let mut prc = Prc::default();
prc.max_parameter = 256;
// Should compile but panic.
unsafe {
    do_something_with_prc_config(&prc.assume_verified());
}

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.