flacenc
Function encode_with_fixed_block_size

pub fn encode_with_fixed_block_size<T: Source>(
    config: &Verified<Encoder>,
    src: T,
    block_size: usize,
) -> Result<Stream, EncodeError>

Encodes Source to Stream.

This is the main entry point of this library crate.

Errors

This function returns EncodeError that contains a SourceError when it failed to read samples from src.

Panics

This function panics only by an internal error.

Examples

use flacenc::config;
use flacenc::error::Verify;
use flacenc::source::MemSource;

let (signal_len, block_size, channels, sample_rate) = (32000, 160, 2, 16000);
let signal = vec![0i32; signal_len * channels];
let bits_per_sample = 16;
let source = MemSource::from_samples(&signal, channels, bits_per_sample, sample_rate);
let result = encode_with_fixed_block_size(
    &config::Encoder::default().into_verified().unwrap(),
    source,
    block_size
);
assert!(result.is_ok());