flacenc::source
Trait Fill

pub trait Fill {
    // Required methods
    fn fill_interleaved(
        &mut self,
        interleaved: &[i32],
    ) -> Result<(), SourceError>;
    fn fill_le_bytes(
        &mut self,
        bytes: &[u8],
        bytes_per_sample: usize,
    ) -> Result<(), SourceError>;
}

Traits for buffer-like objects that can be filled by Source.

An implementation of Source::read_samples is expected to call one of the fill_* method declared in this trait.

Note that arguments of Fill can be shorter than the actual block size (or FrameBuf size). An impl of Fill must accept the samples that is shorter than the pre-defined length. On the other hand, Fill is expected to return an error if the number of samples is larger than the block size.

Required Methods

fn fill_interleaved(&mut self, interleaved: &[i32]) -> Result<(), SourceError>

Fills the target variable with the given interleaved samples.

Errors

This may fail when configuration of Fill is not consistent withto the input interleaved values.

Examples

FrameBuf implements Fill.

let mut fb = FrameBuf::with_size(8, 1024).unwrap();
fb.fill_interleaved(&[0i32; 8 * 1024]);

fn fill_le_bytes( &mut self, bytes: &[u8], bytes_per_sample: usize, ) -> Result<(), SourceError>

Fills target with the little-endian bytes that represent samples.

Errors

This may fail when configuration of Fill is not consistent with the input bytes or bytes_per_sample values.

Examples

FrameBuf implements Fill.

let mut fb = FrameBuf::with_size(2, 64).unwrap();
// Note that `FrameBuf` (or `Fill` in general) accepts shorter inputs.
fb.fill_le_bytes(&[0x12, 0x34, 0x54, 0x76, 0x56, 0x78, 0x10, 0x32], 2);
// this FrameBuf now has 2 channels with elements:
//   - channel-1 (left) : [0x3412, 0x7856]
//   - channel-2 (right): [0x7654, 0x3210]