flacenc::source
Trait Seekable

pub trait Seekable: Source {
    // Required methods
    fn len(&self) -> usize;
    fn read_samples_from<F: Fill>(
        &mut self,
        offset: usize,
        block_size: usize,
        context: &mut F,
    ) -> Result<usize, SourceError>;

    // Provided method
    fn is_empty(&self) -> bool { ... }
}

Trait representing seekable variant of Source.

This trait is not currently used in the encoder, but some encoding algorithm in future may require that a source is seekable.

Required Methods

fn len(&self) -> usize

Returns the length in per-channel samples

fn read_samples_from<F: Fill>( &mut self, offset: usize, block_size: usize, context: &mut F, ) -> Result<usize, SourceError>

Seeks to the specified offset from the beginning.

Errors

This function can return SourceError when read is failed.

Provided Methods

fn is_empty(&self) -> bool

Returns true if the source contains no samples.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.